@apepper 2019-06-12
What is the state of this PR, @trueadm and @gaearon? It is now half a year old and still looks promising.

I would really welcome if this patch or a similar patch would land in react.