@trueadm 2020-01-23
I don't think we plan on revisiting this. There's a risk that changing this could impact/affect the quality and correctness in the escaping of characters. Furthermore, the benchmark figures above are no longer accurate and the difference is much less noticeable compared to before (presumably due to RegEx perf improvements in V8.